refactor: separate login steps
